src: cmd_interface: smarter host/service search
Previously, to indicate a host/service, you needed to type it exactly the way it was registered on Nagios (along with the exact case format).
Thanks to the status.dat file registration, we can map all the hosts and services checked by Nagios and - this way - compare any host/service typed by the user with the effectively registered one.
This comparison is explicitly done in a function that now checks everything regardless of string case formats. Also this allowed us to check host/service by typing only part of them (first part).
For example, if you have an host (or a service, they're checked the same way) called "Windows-7", you'll be allowed to indicate it in any of the following ways:
 - "Windows-7"
 - "windows-7"
 - "Wind"
 - "windo"
 - ...

Obviously, we assume there's no other hosts whose name is starting with "Wind"/"Windo".